requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
astronomer-cosmos[dbt-bigquery]
protobuf
apache-airflow-providers-google
pandas
sqlite-utils
google-cloud-bigquery-storage

astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy